+++ anymotion_sdk/client.py
@@ -1,8 +1,10 @@
 import os
 import time
+import warnings
 from collections import namedtuple
 from logging import getLogger
 from pathlib import Path
+from tempfile import TemporaryDirectory
 from typing import Dict, List, Optional, Union
 from urllib.parse import urljoin, urlparse, urlunparse
 
@@ -68,11 +70,11 @@ class Client(object):
         self.session = session
 
         if client_id is None or client_id == "":
-            raise ClientValueError(f"Client ID is not set.")
+            raise ClientValueError("Client ID is not set.")
         self.client_id = client_id
 
         if client_secret is None or client_secret == "":
-            raise ClientValueError(f"Client Secret is not set.")
+            raise ClientValueError("Client Secret is not set.")
         self.client_secret = client_secret
 
         parts = urlparse(api_url)
@@ -229,8 +231,8 @@ class Client(object):
         path: Optional[Union[str, Path]] = None,
         exist_ok: bool = False,
         fix_suffix: bool = False,
-    ) -> None:
-        """Download file from drawing_id.
+    ) -> Path:
+        """Download a file from drawing_id.
 
         Args:
             drawing_id
@@ -238,6 +240,9 @@ class Client(object):
             exist_ok: if false (default), FileExistsError is raised if the target file
                 already exists.
             fix_suffix: If the extension of path is invalid, correct it.
+
+        Returns:
+            The path to the downloaded file.
 
         Raises:
             FileExistsError
@@ -272,6 +277,30 @@ class Client(object):
         with path.open("wb") as f:
             f.write(response.raw.content)
         logger.info(f"Download file to {path}.")
+
+        return path
+
+    def download_and_read(self, drawing_id: int):
+        """Download and read a file from drawing_id."""
+        try:
+            from .extras import read_image, read_video
+        except ImportError:
+            warnings.warn(
+                "The extras package is not installed. "
+                "Install as follows: pip install anymotion-sdk[cv]"
+            )
+            logger.warning(
+                "Skip the download_and_read method "
+                "because the extras package is not installed."
+            )
+            return
+
+        with TemporaryDirectory() as dir_path:
+            file_path = self.download(drawing_id, path=dir_path)
+            if get_media_type(file_path) == "image":
+                return read_image(file_path)
+            else:
+                return read_video(file_path)
 
     def extract_keypoint(
         self,
@@ -319,7 +348,7 @@ class Client(object):
         Raises:
             RequestsError: HTTP request fails.
         """
-        url = urljoin(self._api_url, f"drawings/")
+        url = urljoin(self._api_url, "drawings/")
         json: Dict[str, Union[int, list, dict]] = {"keypoint_id": keypoint_id}
         if rule is not None:
             json["rule"] = rule
@@ -335,7 +364,7 @@ class Client(object):
         Raises:
             RequestsError: HTTP request fails.
         """
-        url = urljoin(self._api_url, f"analyses/")
+        url = urljoin(self._api_url, "analyses/")
         json: Dict[str, Union[int, list, dict]] = {"keypoint_id": keypoint_id}
         if rule is not None:
             json["rule"] = rule